requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
idna==3.10
jinja2==3.1.6
markupsafe==3.0.2
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.11.1
pydantic-core==2.33.0
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.1
typing-extensions==4.13.0
typing-inspection==0.4.0
uvicorn==0.34.0
itsdangerous==2.2.0